sol2.cfront: Added (int) cast to fix: 	semantics of "<" change in ANSI C; use explicit cast.

+++ trunk/ptclassic/src/octtools/vem/utility/parse.c
@@ -131,7 +131,7 @@ STR *rtn;			/* Remaining portion of string  */
     
     if (!masterSpec) return VEM_FALSE;
 
-    VEM_CHECK(STRLEN(masterSpec) < MAX_SPEC_SIZE, "Instance spec too big");
+    VEM_CHECK((int)STRLEN(masterSpec) < MAX_SPEC_SIZE, "Instance spec too big");
     STRMOVE(ispec, masterSpec);
     p = ispec;
